refactor player collisions; finish charge atk
remove mirrorx from recent commit since only enemies will use that
functionality (until I can find a workaround for player with their
up/down shenanigans)

+++ Soulslite/Assets/Game/code/systems/CameraSystem.cs
@@ -104,7 +104,7 @@ public class CameraSystem : MonoBehaviour
     /**************************
      *         Shake          *
      **************************/
-    public void ActivateShake(int magnitude, float length)
+    public void ActivateShake(float magnitude, float length)
     {
         shakeAmt = 100 * (magnitude / 100f);
         InvokeRepeating("CameraShake", 0, 0.01f);
